refactor: Extracted search result endpoints alone
